refactor and improve types
