+++ src/types/main.ts
@@ -1,0 +1,1 @@
+export { ModuleOptions, Umami } from './index';
